+++ tests/test_sources.py
@@ -5,6 +5,13 @@ This forces us to update tests on ingest!
 """
 
 import pytest
+from sqlalchemy import or_
+
+
+def test_sources(db):
+    # Test that Sources has expected number of entries
+    n_sources = db.query(db.Sources).count()
+    assert n_sources == 7, f"found {n_sources} sources"
 
 
 @pytest.mark.parametrize(
@@ -14,3 +21,23 @@ import pytest
 def test_sources_reference(db, reference, value):
     n_sources = db.query(db.Sources).filter(db.Sources.c.reference == reference).count()
     assert n_sources == value, f"found {n_sources} sources for {reference}"
+
+
+def test_coordinates(db):
+    # Verify that all sources have valid coordinates
+    t = (
+        db.query(db.Sources.c.source, db.Sources.c.ra_deg, db.Sources.c.dec_deg)
+        .filter(
+            or_(
+                db.Sources.c.ra_deg.is_(None),
+                db.Sources.c.ra_deg < 0,
+                db.Sources.c.ra_deg > 360,
+                db.Sources.c.dec_deg.is_(None),
+                db.Sources.c.dec_deg < -90,
+                db.Sources.c.dec_deg > 90,
+            )
+        )
+        .astropy()
+    )
+
+    assert len(t) == 0, f"{len(t)} Sources failed coordinate checks: {t}"
